cyclades: Add VolumeType DB model
Create new 'VolumeType' database model to represent the available types
for Cyclades Volumes. Currently, this model contains only the volume's
disk template and the human-readable name, but can be extended
to include extra specs, e.g. QoS specs.

The 'disk_template' column from the 'Flavor' and 'Volume' models is
replaced with a foreign key to the 'VolumeType' model.