deploy: Fix status in case of parallel context
Fabric's @parallel creates separate processes in order to run the
requested tasks in parallel. Thus the CofigParser object held in memory
(status.cfg) is not the same across different processes.

To ensure validity, any read/update/write
operation must be done by using the status file in permanent storage
and not the status.cfg object.

For atomicity this must be done after acquiring the corresponding
lock. Introduce @_lock_read_write for this purpose.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>