snf_django: Fix unicode error in proxy requests
Pithos and Cyclades services need to proxy Astakos API calls for enabling
the web clients to access them.

Proxy requests containing non ASCII characters used to fail.

The reason was that the proxy tried to make a request using the original
request path.
However Django forces the request path to be unicode.
Therefore, the httplib used to fail because it tried to concatenate unicode
and bytestrings.

This fix changes the original implementation and from now on it sends
the requested URL encoded in UTF-8.