deploy: Do create a VG if not extra disk found
If an extra disk found (node_info.extra_disk) then use it.
Otherwise create a raw file under / and losetup it.

Add a new option `vg_size` to reflect the size of the volume group
(the physical group actually) of a Ganeti node. Use it to
create the raw file if the extra disk is not found.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>