deploy/ci: Fix package installation process
When local packages are used (use_local_packages = True in
deploy.conf) we do:

 dpkg -i <local_package.deb> || apt-get install -f

The first command leaves a broken package while the latter fixes it.
For some reason this DOES upgrade the installed packages that the
broken one depends on. To bypass this strange behavior,
after dpkg -i we run apt-mark hold.

Additionally remove any dependencies in all components in order
to check if control files are generated properly. Leave only
python-svipc for ArchipSynnefo and memcached in Cyclades.

Install snf-image 0.14 that does still work with nfs and does not
require a rados backend.

Rename WEB to Webproject.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>