astakos: Fix stats for new projects
The recent changes in the Astakos projects implementation changed the
contents of the 'quotaholder_app_holding' table. This table now contains
holdings for both users and projects. Also, the 'holder' column contains
user or project UUIDs, prefixed by 'user:' or 'project:' respectively.
This commit updates the stats mechanism in order to reflect the
aforementioned changes.