cyclades: Accept an empty image when using volumes
Accept an empty 'imageRef' attribute when creating a server with
volumes. The attribute must always be present even with an empty value.
In case the server's image is left blank, we use the source
(image/snapshot) of the first volume. Also, if an image is specified
both for the server and for the first volume, we check that the image is
the same.