deploy/ci: Make vcluster use debootstrap
Introduce a new helper script, mkimage.sh that creates a debian base
raw image with debootstrap so that it can be used in a virtual
cluster.

Install network-manager and a hook to update hostname with info
provided by DHCP responce.

Add new conf setting for the size of the image.

Create another empty image as well so that the VM can use it for
LVM.

Remove unused settings.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>