astakos: Create system project on user acceptance
Delay the creation of user's system project until the user is accepted,
to avoid polluting the database with unneeded project objects.

There is a slight chance that a user UUID (to be used for a system project)
collides with an existing project UUID. In this case, one has to delete the
user and make a new one in its place.