pithos: Change default disposition type to inline
Since version 0.15 a specific content disposition type can be enforced in
requests for an object's content, using the 'disposition-type' optional
parameter.

If no disposition-type is specifically requested or an invalid value is passed,
the default disposition type used to be 'attachment'.

This commit changes the default disposition type to 'inline'.