cyclades: Store image version additional to UUID
Store the image's version together with the image's UUID, in order to
track the exact image (Pithos file) that has been used for a server or
volume. This is required because Pithos files retain their UUID when the
file's data or metadata are updated. The (uuid, version) tuple is
required in order to fully determine the file that has been used.