deploy: Support Ganeti 2.10
..and later but NOT 2.8.

In Ganeti 2.10 --no-lvm-storage and --no-drbd-storage options of
gnt-cluster init are removed in favor of --enabled-disk-templates.

Additionally the packaging has changed. A valid ganeti installation
has: ganeti-haskell-2.10, ganeti-htools-2.10, ganeti-2.10,
snf-ganeti, ganeti2.

Since snf-ganeti (or ganeti) depends on all other we remove
ganeti-haskell and ganeti-htools from the REQUIRED_PACKAGES of
the Ganeti component. ganeti2 is a dummy transitional package that
satisfies external components dependencies (snf-image,
snf-cyclades-gtools, etc.)

The version to be installed is configurable in packages.conf.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>