pithos: Fix wrong quotas when moving objects
Quotas were not accounted correctly when moving an object or a folder to a
path that another object already exists.

This commit fixes this issue.

In addition to this, the backend during move operations used to issue the quota
change for each separate operation (copy & delete) and for each affected object
(in the case of bulk move operations).

This commit introduces the following enhancement:

In case of a move operation to the same account and to the same project
computes the whole quota change and contacts the quota holder component only
once (at the end of the whole procedure) for the specific account/project.

However, we do not perform the same for such copy operations because we want to
catch quickly failures due to quota restrictions.

Fixes #2