admin-ui: Update layout of table and add "Clear All" btn
There are now checkboxes in the 1st column of the main table and in
the 1st column of the table of selected items. The checkboxes in
the main table are checked when a row is selected and unchecked
when o row is not selected.
The checkboxes in the table of selected items are always checked.

The "Clear All" button triggers a confirmation modal and then the user
can choose to deselect all his selections (not only the visible).