cyclades: Fix backend logic wrt ConnectNetwork
Since Ganeti 2.10, networks are connected to nodegroups with mode,
link and vlan info.

As a result group_list in GetNetwork() response include four fields
per group and not three. Fix reconciliation to be aware of it.

Additionally the OpNetworkConnect() takes an extra parameter;
network_vlan. This currently is not optional. Therefore let
ConnectNetwork() RAPI method pass an empty string as vlan info.

This will change as soon as synnefo will support private networks on
top of openvswitch.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>