plankton: Delay creation of Pithos backend pool
Delay the creation of the Pithos backend pool until an object from the
pool is actually needed. Until now, the pool was created when the
'synnefo.plankton.backend' module was imported. This caused an issue
to 'snf-dispatcher' which performs a 'fork()' in order to become a
daemon and 'objpool' which checks that the pid of the process who
creates and uses the pool is the same.