admin: Load separate JS files for each view type
Make the page more responsive by loading only the JS files needed. To do
this, we take advantage of the {% block custom-javascript %} that is
defined in base.html.

Note that common JS files such as jQuery and Bootstrap are defined in
base.html and are thus loaded in every page.