pithos: Backend refactoring
The implementation for bulk deleting/copying/moving paths starting with
the provided path + the delimiter argument (if such is provided) used to
query the database for the specific paths and perform in a loop the same
actions performed for the single delete/copy/move operations.
This commit substitutes the code inside the loop by calling the
respective method for the derived path (using recursion).