admin: Improve logging
This commit does the following:

a) Removes all unnecessary logging calls.
b) Adds a logging mechanism for post-mortem analysis.

The new logging mechanism, accessible from admin.utils.admin_log, logs
by default logs the name of the user and the caller view and allows the
user to add useful info such as the arguments for this view.