admin: Fix auth provider filters
Due to a bug in Django (https://code.djangoproject.com/ticket/14645), it
is not possible to use the "HAS NOT" filter for auth providers using the
`exclude` method. Fix this behavior using a subquery to filter the users
that have the requested auth provider and then exclude them.

Fix apyrgio/synnefo#199