pithos: Restore XML input in uploads using hashmaps
In version 0.15 the option to create an object by providing the object's
hashmap in XML format, instead of the actual data, was disabled. Restore
this operation by substituting the 'minidom' module (vulnerable to some
security implications) with the 'lxml' module for performing the XML
parsing. Also fix the respective test case.

Close grnet/synnefo#155