admin: Show project quotas
Project quotas are shown in two different places:

a) In the project summary, where they have to be displayed horizontally
and
b) in the details of a project, where they should be shown vertically.

Besides showing the project quotas properly, we must show and the
requested project application resources properly, since we currently
have no direct way to show them.

Therefore, for each project we will show either the latest state of its
quotas or the latest state of its application resources.

Fix apyrgio/synnefo#215