admin: Add tests for views
The tests that are added are both unit and integration tests and are a
first step towards testing all the security aspects of admin.

Note that some of the tests come from the helpdesk app, and are stored
as tests_old.py for reference reasons. In the future, the tests_old.py
file will be removed.