admin: Show IP history only for floating IPs
By default, the IP log contains the history of all IPs, both floating
and private. However, showing the IP history for private IPs has two
serious drawbacks:

1) Address class between floating-private IPs.

There are cases when some floating IPs and public IPs have the same
address name. In this case, it is not easy to the history only for the
private or floating IP, since the IP type is not registered in the log.
What's registered however is a network id from which we can extract this
information. However, this means that in order to get details for an IP,
we need to have its latest network ID, which may not be available for
old entries in the IP log and further complicates things.

2) Private IPs are one-use only.

A private IP cannot be detached from a VM and be given in another VM, at
least not in the same way as floating IPs are detached. A private IP
lives as long as a VM is connected to a private network, and then gets
destroyed. This info is not very useful.