admin: Work-around a weird behavior of Django's QS
In the filter functions of project, there is a `filter_user` function
that has very weird results when it gets passed to `values_list`. I have
checked that depending on whether `values_list` is in the provided
queryset or not, the returned results may omit the base project of the
user.

This is countered by using `select_related` on the `owner__uuid` field
of the Project model and then evaluating the queryset before using
`values_list` on it.

I have found no justification for this behavior. I believe that what we
do here lies on the gray area of Django ORM since we use `values_list`
for attributes such as "owner__uuid" which may not exist in the Project
instance. It should be interesting however to report it to the Django
team, once I can reproduce it somewhere else.