snf_django: Add database router for Synnefo apps
Add Django database router to support using Synnefo with multiple
databases. In a gist, if the administrator has added in the DATABASES
setting a database with the "cyclades" or "astakos" alias, and
SynnefoRouter is defined in DATABASE_ROUTERS, then that database will be
used to read/write the respective Cyclades or Astakos models.