admin: Add setting for enabling charts
Add a setting (ADMIN_ENABLE_CHARTS) for enabling the charts presentation
of Astakos/Cyclades resources. This setting will default to False.

The charting software that is used is Highcharts by HighSoft AS. It is
due to the licensing nature of Highcharts that charts are disabled by
default. You can view the Highcharts license [1] and the accompanying
FAQ [2], to decide whether to enable it or not.

[1]: www.highcharts.com/license, http://shop.highsoft.com/highcharts.html
[2]: http://shop.highsoft.com/highcharts.html,
     http://shop.highsoft.com/faq/non-commercial