admin: Update template tags
Add Django template tag that returns the filter type for a given filter
and add a choice template filter that returns a list with the available
choices for a filter (only for MultipleChoiceFilter types).

Also, use an array of filters instead of generator in the templatetags
context, since the latter cannot be re-iterated.