Fix burnin bug: Index error in pithos account_get
Bug description: pithos test assumed there are always some extra
containers in the tested pithos account and checked for their
existence, causing an Index error to be raised when starting the
test with an empty account.

Solution: Do not check for containers outside the ones created
specifically for the test.