burnin: Don't attach two NICs to public networks
Burnin used to attach two NICs to public networks and test that the
virtual machine is accessible through both NICs. This test used to fail
because the virtual machine has only one gateway and the firewall
doesn't allow traffic that comes from one NICs to be routed through the
other.

Remove the above test and instead first detach the virtual machine from
the public network and then re-attach it (hence test hotplugging
functionality).