cyclades: Fix snapshots of attached volumes
Fix major bug when taking snapshots of non-root volumes. The custom
'SnapshotInstance' method of 'GanetiRAPIClient' always took a snapshot
of the root volume (index=0) of an instance, ignoring the volume that
the user requested. This resulted in always taking a snapshot of the
root volume which in many cases did not fail.