cyclades: Update snapshot based on Ganeti job
Make Cyclades update the state of a snapshot in Pithos DB based on
the result of the corresponding Ganeti job. This is required to be
performed by Cyclades, because Pithos has no way to know whether an
unavailable snapshot will ever by completed or not (e.g. the Ganeti
job failed or it still has not completed).

Refs grnet/synnefo#231