cyclades: Reconciliation for stale snapshots
Extend 'reconcile-servers' command to perform reconciliation for
snapshots, which can occur if the Ganeti job failed but the
corresponding message from eventd has never been processed. In order to
make reconciliation possible, Cyclades store a special metadata on the
snapshot that contains the Ganeti ID and the job ID that will create the
snapshot.

Closes grnet/synnefo#231