admin: Fix bug in default view calculation
If ADMIN_VIEWS_ORDER had as first element a non-existent view, then this
would be returned as the default view, despite the fact that the
ADMIN_VIEWS dictionary would not include it. Fix this by consulting only
the ADMIN_VIEWS dictionary and update the tests to catch this later on.

Also, use SnfOrderedDict for the creation of ADMIN_VIEWS.