cyclades: Fix bug in DB migration
Fix bug in migration 0102 which saves in the 'subnet_ids' column of the
network object the IDs of all subnets that are associated with this
network. A typo in this migration resulted in the 'subnet_ids' field to
contain the id of the network. Also, provide a DB migration to fix this
issue for those that have already installed 0.16rc1.