admin-ui: Fix bugs regarding arrow-buttons in tables
In some cases, the arrow button that is included in table rows was
not clickable. That was happening because the summary area was
expanding over the arrow.

With the addition of z-index property this problem fixed.

Also, in some cases, there was a shake when summary area was
expanding. This occurred from measuring the height of the span in
the arrow-button instead of the button's height.