admin: Return QuerySet when a filter has no input
If a filter has not been given an input, then django-filter should
normally not call its function. However, it seems this is not the case
with the index Volume filter, whose function is called with no value and
as a result, it returns an empty QuerySet.

Fix this issue by checking in filter_index if it has been passed with
any parameters. Also, since this issue may appear in other filter
functions too, add the same check in every function that doesn't have
it.

NOTE: Model filters have a decorator that automatically does this check
for them.