astakos: Add email notifications recipients
Astakos accepts three lists of people who receive application
notifications, namely 'ADMINS', 'HELPDESK' and 'MANAGERS'. Then it sends
email notifications in a preconfigured way in these lists (ie. project
creation notifications are always sent to the 'HELPDESK' and 'MANAGERS'
email lists).

We would like to enable Astakos to specify explicitly the lists in which
every notification will be sent. This commit adds the corresponding
settings that will enable this feature.

For convenience, Astakos groups the notifications in three categories
and let the user define the recipients for these.
  - ACCOUNT_NOTIFICATIONS_RECIPIENTS receive notifications for 'account
    pending moderation' and 'account activated' actions.
  - FEEDBACK_NOTIFICATIONS_RECIPIENTS receive feedback notifications
  - PROJECT_NOTIFICATIONS_RECIPIENTS receive notifications for 'project
    creations' and 'project modification' actions.

Furthermore, one can explicitly specify the recipients for a specific
notification by using one of the following settings:
  - ACCOUNT_PENDING_MODERATION_RECIPIENTS
  - ACCOUNT_ACTIVATED_RECIPIENTS
  - PROJECT_CREATION_RECIPIENTS
  - PROJECT_MODIFICATION_RECIPIENTS
By default, these settings are not exposed to the config file.