pithos: Slight change in get_domain_objects
This method gets two keyword arguments: user and check_permissions.
If check_permissions is True returns only the object accessible to the user.
Otherwise list all the objects for the specific domain ignoring
permissions.
The latter case is only meaningful if user is not provided,
therefore in that case it raises AssertionError.