ci: Resource project handling
ci user may set `Deployment.project` setting to force client to
create resources bound to the specified project. Additionally to ease up ci
runs for users with limited quota on system project, the client detects
a project suitable to fit the underlying resource and use it to create
the resource.