cyclades: Fix wait-for-sync on stopped instances
Pass 'wait_for_sync' option only on started Ganeti instances, because Ganeti
will fail with 'OpPrereqError' in case the 'wait_for_sync' is set to
False and the instance has deactivated disks.

Note that Cyclades will not let a user neither attach a volume to an instance
that is is in transition state, e.g. stopping, nor change the state of
an instance if there is a pending volume attachment. However, in case
the administrator manually modifies the instance via Ganeti, there is a
race between checking the state of the instance and the time that Ganeti
job will run that cannot be prevented.

Closes grnet/synnefo#314