deploy: Fix GanetiDev component
This component is supposed to clone a ganeti repo and setup
a Ganeti QA cluster.

We provide a qa-init.sh that installs all required packages (apt-get
install and cabal install) and prepares the node for QA.

We provide a configure-ganeti similar to the one the debian packages
use so make install && /etc/init.d/ganeti restart will override
the existing binaries.

To install ganeti from source run on all nodes:

 # cd /srv/srv/ganeti
 # ./qa-init.sh
 # ./configure-ganeti
 # make && make install
 # cp doc/examples/ganeti.initd /etc/init.d/ganeti
 # /etc/init.d/ganeti restart

To run the QA do:

 # cd /srv/src/ganeti/qa
 # PYTHONPATH=/srv/src/ganeti/ ./ganeti-qa.py --yes-do-it /root/qa-sample.json

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>