ui: Fix check about which images support ssh keys
Fix the check about which images support ssh keys. Besides the operating
system of the image we must also check if the properties of the image
specify that the 'EnforcePersonality' task or all tasks will be
excluded. Also, if the image has no properties, then that it is
equivalent with exclude all tasks.