astakos: Create LDAP authentication provider
Create the 'LDAPAuthProvider' to authenticate against an LDAP service.
This provider is making use of the 'django-auth-ldap' package. However,
the 'django_auth_ldap.backend.LDAPBackend' is make to work as a Django
authentication backend, and so to automatically creates a User object.
Because this does not match with how Astakos is handling authentication
of third party providers, we override the 'get_or_create_user' method to
create a mocked user object, which will be used by our provider to
create the final AstakosUser object.