ui: User defined image sections
permit custom defined image sections in image selection view of the vm
create wizard. The value of the `LISTING_SECTION` is used as the name of the
section. To avoid abuse of the feature only images owned by user uuids
included in UI_IMAGE_LISTING_USERS setting will be given extra section
handling.