cyclades: Add `origin_size` in Volume fields
Add the `origin_size` parameter as a field of the Volume model. This
parameter is passed during volume creation to Ganeti and is used only
for that operation. Also, create a schema migration for this field.

There is no need for data migration for this field since if a volume has
already been created, `origin_size` is no longer necessary and we can
carry on using the `size` parameter.