admin-ui: Execute filtering when input value changes
This modification affects the text inputs that are used for filtering
the content of the tables.
In the past when the user was trying to paste some text in the input,
without using a keyboard shortcut, he had to press a key except enter
or space, in order to trigger the filtering function.
Now, when he pastes some text, the filtering is executed immediately.